Move CompletedPuzzle comp into game/
